# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/notifications": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Create Notification",
                "parameters": [
                    {
                        "description": "Create Notification",
                        "name": "notification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notification_dto.CreateNotificationDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/ztm/stop/{stopId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ztm"
                ],
                "summary": "Get ZTM stop with all line numbers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ZTM Stop ID",
                        "name": "stopId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
        "notification_dto.CreateNotificationDto": {
            "type": "object",
            "required": [
                "deviceToken",
                "lineNumber",
                "stopId",
                "stopName"
            ],
            "properties": {
                "deviceToken": {
                    "type": "string",
                    "example": "fIuoGe66REq_eyZaN2V8E0"
                },
                "lineNumber": {
                    "type": "string",
                    "example": "199"
                },
                "providerName": {
                    "type": "string",
                    "example": "ztm"
                },
                "reminderTime": {
                    "type": "integer",
                    "maximum": 60,
                    "minimum": 1,
                    "example": 10
                },
                "stopId": {
                    "type": "string",
                    "example": "1461"
                },
                "stopName": {
                    "type": "string",
                    "example": "Przymorze Wielkie"
                }
            }
        }
    }
}